# scripts/sandbox_openapi_base.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Treadstone Sandbox",
    "description": "\nTreadstone browser and shell runtime.\n\n- Browser CDP: `/cdp/json/version`\n- Interactive shell: `/terminal`\n- Browser, shell and file tools over Streamable HTTP: `/mcp`\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/shell/exec": {
      "post": {
        "tags": [
          "shell"
        ],
        "summary": "Exec Command",
        "description": "Execute command in the specified shell session\nSupports SSE streaming if Accept header contains 'text/event-stream'",
        "operationId": "exec_command_v1_shell_exec_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShellExecRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response_ShellCommandResult_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-fern-sdk-group-name": "shell",
        "x-fern-sdk-method-name": "exec_command"
      }
    },
...
    "/v1/file/read": {
      "post": {
        "tags": [
          "file"
        ],
        "summary": "Read File",
        "description": "Read file content",
        "operationId": "read_file",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FileReadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response_FileReadResult_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-fern-sdk-group-name": "file",
        "x-fern-sdk-method-name": "read_file"
      }
    },
    "/v1/file/write": {
      "post": {
        "tags": [
          "file"
        ],
        "summary": "Write File",
        "description": "Write file content (supports both text and binary files)\n\nFor binary files, set encoding to 'base64' and provide base64-encoded content.\nFor text files, use default 'utf-8' encoding.",
        "operationId": "write_file",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FileWriteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response_FileWriteResult_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-fern-sdk-group-name": "file",
        "x-fern-sdk-method-name": "write_file"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BashCommandStatus": {
        "type": "string",
        "enum": [
          "running",
          "completed",
          "no_change_timeout",
          "hard_timeout",
          "terminated"
        ],
        "title": "BashCommandStatus",
        "description": "Shell command execution status (compatible with OpenHands)"
      },
...
      "ConsoleRecord": {
        "properties": {
          "ps1": {
            "type": "string",
            "title": "Ps1",
            "description": "Command prompt"
          },
          "command": {
            "type": "string",
            "title": "Command",
            "description": "Executed command"
          },
          "output": {
            "type": "string",
            "title": "Output",
            "description": "Command output",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "ps1",
          "command"
        ],
        "title": "ConsoleRecord",
        "description": "Shell command console record model"
      },
...
      "FileContentEncoding": {
        "type": "string",
        "enum": [
          "utf-8",
          "base64",
          "raw"
        ],
        "title": "FileContentEncoding",
        "description": "File content encoding type"
      },
...
      "FileReadRequest": {
        "properties": {
          "file": {
            "type": "string",
            "title": "File",
            "description": "Absolute file path"
          },
          "start_line": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Line",
            "description": "Start line (0-based)"
          },
          "end_line": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Line",
            "description": "End line (not inclusive)"
          },
          "sudo": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sudo",
            "description": "Whether to use sudo privileges",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "FileReadRequest",
        "description": "File read request"
      },
      "FileReadResult": {
        "properties": {
          "content": {
            "type": "string",
            "title": "Content",
            "description": "File content"
          },
          "file": {
            "type": "string",
            "title": "File",
            "description": "Path of the read file"
          }
        },
        "type": "object",
        "required": [
          "content",
          "file"
        ],
        "title": "FileReadResult",
        "description": "File read result"
      },
...
      "FileWriteRequest": {
        "properties": {
          "file": {
            "type": "string",
            "title": "File",
            "description": "Absolute file path"
          },
          "content": {
            "type": "string",
            "title": "Content",
            "description": "Content to write (text or base64 encoded for binary)"
          },
          "encoding": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FileContentEncoding"
              },
              {
                "type": "null"
              }
            ],
            "description": "Content encoding: utf-8 for text, base64 for binary data",
            "default": "utf-8"
          },
          "append": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Append",
            "description": "Whether to use append mode",
            "default": false
          },
          "leading_newline": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Leading Newline",
            "description": "Whether to add leading newline (only for text mode)",
            "default": false
          },
          "trailing_newline": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trailing Newline",
            "description": "Whether to add trailing newline (only for text mode)",
            "default": false
          },
          "sudo": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sudo",
            "description": "Whether to use sudo privileges",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "file",
          "content"
        ],
        "title": "FileWriteRequest",
        "description": "File write request - supports both text and binary content"
      },
      "FileWriteResult": {
        "properties": {
          "file": {
            "type": "string",
            "title": "File",
            "description": "Path of the written file"
          },
          "bytes_written": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bytes Written",
            "description": "Number of bytes written"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "FileWriteResult",
        "description": "File write result"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Response_FileReadResult_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Whether the operation was successful",
            "default": true
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "description": "Operation result message",
            "default": "Operation successful"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FileReadResult"
              },
              {
                "type": "null"
              }
            ],
            "description": "Data returned from the operation"
          }
        },
        "type": "object",
        "title": "Response[FileReadResult]"
      },
...
      "Response_FileWriteResult_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Whether the operation was successful",
            "default": true
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "description": "Operation result message",
            "default": "Operation successful"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FileWriteResult"
              },
              {
                "type": "null"
              }
            ],
            "description": "Data returned from the operation"
          }
        },
        "type": "object",
        "title": "Response[FileWriteResult]"
      },
      "Response_ShellCommandResult_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Whether the operation was successful",
            "default": true
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "description": "Operation result message",
            "default": "Operation successful"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ShellCommandResult"
              },
              {
                "type": "null"
              }
            ],
            "description": "Data returned from the operation"
          }
        },
        "type": "object",
        "title": "Response[ShellCommandResult]"
      },
...
      "ShellCommandResult": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id",
            "description": "Shell session ID"
          },
          "command": {
            "type": "string",
            "title": "Command",
            "description": "Executed command"
          },
          "status": {
            "$ref": "#/components/schemas/BashCommandStatus",
            "description": "Command execution status"
          },
          "output": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output",
            "description": "Command execution output, only has value when status is completed"
          },
          "console": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ConsoleRecord"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Console",
            "description": "Console command records"
          },
          "exit_code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exit Code",
            "description": "Command execution exit code, only has value when status is completed"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "command",
          "status"
        ],
        "title": "ShellCommandResult",
        "description": "Shell command execution result model"
      },
...
      "ShellExecRequest": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id",
            "description": "Unique identifier of the target shell session, if not provided, one will be automatically created"
          },
          "exec_dir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exec Dir",
            "description": "Working directory for command execution (must use absolute path)"
          },
          "command": {
            "type": "string",
            "title": "Command",
            "description": "Shell command to execute"
          },
          "async_mode": {
            "type": "boolean",
            "title": "Async Mode",
            "description": "Whether to execute command asynchronously (default: False for async, False for synchronous execution)",
            "default": false
          },
          "timeout": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeout",
            "description": "Maximum time (seconds) to wait for command completion before returning running status"
          }
        },
        "type": "object",
        "required": [
          "command"
        ],
        "title": "ShellExecRequest",
        "description": "Shell command execution request model"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}